add extra page for session data
